refactor. join 메서드 관리 이동room -> roomMember
