+++ src/main/java/com/jungle/studybbitback/domain/room/service/RoomMemberService.java
@@ -2,6 +2,8 @@ package com.jungle.studybbitback.domain.room.service;
 
 import com.jungle.studybbitback.domain.member.entity.Member;
 import com.jungle.studybbitback.domain.member.repository.MemberRepository;
+import com.jungle.studybbitback.domain.room.dto.roommember.JoinRoomMemberRequestDto;
+import com.jungle.studybbitback.domain.room.dto.roommember.JoinRoomMemberResponseDto;
 import com.jungle.studybbitback.domain.room.dto.roommember.InviteRoomMemberRequestDto;
 import com.jungle.studybbitback.domain.room.dto.roommember.InviteRoomMemberResponseDto;
 import com.jungle.studybbitback.domain.room.dto.roommember.LeaveRoomMemberRequestDto;
@@ -10,8 +12,10 @@ import com.jungle.studybbitback.domain.room.entity.Room;
 import com.jungle.studybbitback.domain.room.entity.RoomMember;
 import com.jungle.studybbitback.domain.room.respository.RoomMemberRepository;
 import com.jungle.studybbitback.domain.room.respository.RoomRepository;
+import com.jungle.studybbitback.jwt.dto.CustomUserDetails;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -24,6 +28,33 @@ public class RoomMemberService {
     private final RoomMemberRepository roomMemberRepository;
     private final RoomRepository roomRepository;
     private final MemberRepository memberRepository;
+
+    //스터디룸 참여(가입)
+    @Transactional
+    public JoinRoomMemberResponseDto joinRoom(Long roomId, JoinRoomMemberRequestDto requestDto) {
+//        Long roomId = requestDto.getRoomId();
+
+        // 로그인한 사용자 정보 가져오기
+        CustomUserDetails userDetails = (CustomUserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+        Long memberId = userDetails.getMemberId();
+
+        Room room = roomRepository.findById(roomId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 방입니다."));
+        Member member = memberRepository.findById(memberId)
+                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 회원입니다."));
+
+        // 이미 방에 참여한 경우 확인
+        if (roomMemberRepository.existsByRoomAndMember(room, member)) {
+            throw new IllegalStateException("이미 참여한 방입니다.");
+        }
+
+        // 방 참여 처리
+        RoomMember roomMember = new RoomMember(room, member);
+        roomMemberRepository.save(roomMember);
+
+        int participantCount = roomMemberRepository.countByRoom(room);
+        return new JoinRoomMemberResponseDto(roomId, memberId, participantCount, memberId +"님이 방에 참여했습니다.");
+    }
 
     //스터디룸 초대
     @Transactional
